[Optimize] optimize the structure of OALRuntime and skywalking-on-compiling-plugin. Optimize comments and project structure.
