--- distribution/graal/skywalking-providers-graal-native/src/main/java/org/apache/skywalking/oap/log/analyzer/provider/graal/LogAnalyzerModuleProviderGraal.java
@@ -1,40 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
-
-package org.apache.skywalking.oap.log.analyzer.provider.graal;
-
-import org.apache.skywalking.oap.log.analyzer.provider.LogAnalyzerModuleProvider;
-import org.apache.skywalking.oap.server.library.module.ModuleStartException;
-import org.apache.skywalking.oap.server.library.module.ServiceNotProvidedException;
-
-/**
- * The old logic of this class used Groovy;
- * however, GraalVM does not support it well, so it has been replaced with an empty implementation.
- * In the future, other methods will be used to implement this class.
- */
-public class LogAnalyzerModuleProviderGraal extends LogAnalyzerModuleProvider {
-    @Override
-    public String name() {
-        return "default-graalvm";
-    }
-
-    @Override
-    public void start() throws ServiceNotProvidedException, ModuleStartException {
-
-    }
-}
